tests: Move publishing tests to `krate::publish` module
