+++ src/tests/http-data/krate_publish_good_badges
