+++ src/tests/http-data/krate_publish_ignored_badges
