+++ src/tests/http-data/krate_publish_new_crate_allow_empty_alternative_registry_dependency
