+++ src/tests/http-data/krate_publish_new_krate_with_dependency
