+++ src/tests/http-data/krate_publish_new_with_renamed_dependency
