+++ src/tests/http-data/krate_publish_publish_after_removing_documentation
