+++ src/tests/krate/mod.rs
@@ -1,27 +1,21 @@
 use crate::{
-    builders::{CrateBuilder, DependencyBuilder, PublishBuilder, VersionBuilder},
-    new_category, new_dependency, new_user, CrateMeta, CrateResponse, GoodCrate, OkBool,
-    RequestHelper, TestApp,
+    builders::{CrateBuilder, PublishBuilder, VersionBuilder},
+    new_category, new_dependency, new_user, CrateMeta, OkBool, RequestHelper, TestApp,
 };
 use cargo_registry::{
-    models::{krate::MAX_NAME_LENGTH, Category},
-    schema::{api_tokens, crates, emails, metadata, versions, versions_published_by},
+    models::Category,
+    schema::{crates, metadata, versions},
     views::{
         EncodableCategory, EncodableCrate, EncodableDependency, EncodableKeyword, EncodableVersion,
         EncodableVersionDownload,
     },
 };
-use std::{
-    collections::HashMap,
-    io::{self, prelude::*},
-    thread,
-    time::Duration,
-};
 
 use chrono::Utc;
 use conduit::StatusCode;
 use diesel::{dsl::*, prelude::*, update};
-use flate2::{write::GzEncoder, Compression};
+
+mod publish;
 
 #[derive(Deserialize)]
 struct VersionsList {
@@ -696,624 +690,6 @@ fn versions() {
 }
 
 #[test]
-fn uploading_new_version_touches_crate() {
-    use diesel::dsl::*;
-
-    let (app, _, user) = TestApp::full().with_user();
-
-    let crate_to_publish = PublishBuilder::new("foo_versions_updated_at").version("1.0.0");
-    user.enqueue_publish(crate_to_publish).good();
-
-    app.db(|conn| {
-        diesel::update(crates::table)
-            .set(crates::updated_at.eq(crates::updated_at - 1.hour()))
-            .execute(&*conn)
-            .unwrap();
-    });
-
-    let json: CrateResponse = user.show_crate("foo_versions_updated_at");
-    let updated_at_before = json.krate.updated_at;
-
-    let crate_to_publish = PublishBuilder::new("foo_versions_updated_at").version("2.0.0");
-    user.enqueue_publish(crate_to_publish).good();
-
-    let json: CrateResponse = user.show_crate("foo_versions_updated_at");
-    let updated_at_after = json.krate.updated_at;
-
-    assert_ne!(updated_at_before, updated_at_after);
-}
-
-#[test]
-fn new_wrong_token() {
-    let (app, anon, _, token) = TestApp::init().with_token();
-
-    // Try to publish without a token
-    let crate_to_publish = PublishBuilder::new("foo");
-    let json = anon
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::FORBIDDEN);
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("must be logged in to perform that action"),
-        "{:?}",
-        json.errors
-    );
-
-    // Try to publish with the wrong token (by changing the token in the database)
-    app.db(|conn| {
-        diesel::update(api_tokens::table)
-            .set(api_tokens::token.eq(b"bad" as &[u8]))
-            .execute(conn)
-            .unwrap();
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo");
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::FORBIDDEN);
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("must be logged in to perform that action"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn invalid_names() {
-    let (_, _, _, token) = TestApp::init().with_token();
-
-    let bad_name = |name: &str, error_message: &str| {
-        let crate_to_publish = PublishBuilder::new(name).version("1.0.0");
-        let json = token
-            .enqueue_publish(crate_to_publish)
-            .bad_with_status(StatusCode::OK);
-
-        assert!(
-            json.errors[0].detail.contains(error_message,),
-            "{:?}",
-            json.errors
-        );
-    };
-
-    let error_message = "expected a valid crate name";
-    bad_name("", error_message);
-    bad_name("foo bar", error_message);
-    bad_name(&"a".repeat(MAX_NAME_LENGTH + 1), error_message);
-    bad_name("snow☃", error_message);
-    bad_name("áccênts", error_message);
-
-    let error_message = "cannot upload a crate with a reserved name";
-    bad_name("std", error_message);
-    bad_name("STD", error_message);
-    bad_name("compiler-rt", error_message);
-    bad_name("compiler_rt", error_message);
-    bad_name("coMpiLer_Rt", error_message);
-}
-
-#[test]
-fn new_krate() {
-    let (_, _, user) = TestApp::full().with_user();
-
-    let crate_to_publish = PublishBuilder::new("foo_new").version("1.0.0");
-    let json: GoodCrate = user.enqueue_publish(crate_to_publish).good();
-
-    assert_eq!(json.krate.name, "foo_new");
-    assert_eq!(json.krate.max_version, "1.0.0");
-}
-
-#[test]
-fn new_krate_with_token() {
-    let (_, _, _, token) = TestApp::full().with_token();
-
-    let crate_to_publish = PublishBuilder::new("foo_new").version("1.0.0");
-    let json: GoodCrate = token.enqueue_publish(crate_to_publish).good();
-
-    assert_eq!(json.krate.name, "foo_new");
-    assert_eq!(json.krate.max_version, "1.0.0");
-}
-
-#[test]
-fn new_krate_weird_version() {
-    let (_, _, _, token) = TestApp::full().with_token();
-
-    let crate_to_publish = PublishBuilder::new("foo_weird").version("0.0.0-pre");
-    let json: GoodCrate = token.enqueue_publish(crate_to_publish).good();
-
-    assert_eq!(json.krate.name, "foo_weird");
-    assert_eq!(json.krate.max_version, "0.0.0-pre");
-}
-
-#[test]
-fn new_with_renamed_dependency() {
-    let (app, _, user, token) = TestApp::full().with_token();
-
-    app.db(|conn| {
-        // Insert a crate directly into the database so that new-krate can depend on it
-        CrateBuilder::new("package-name", user.as_model().id).expect_build(conn);
-    });
-
-    let dependency = DependencyBuilder::new("package-name").rename("my-name");
-
-    let crate_to_publish = PublishBuilder::new("new-krate")
-        .version("1.0.0")
-        .dependency(dependency);
-    token.enqueue_publish(crate_to_publish).good();
-    app.run_pending_background_jobs();
-
-    let crates = app.crates_from_index_head("ne/w-/new-krate");
-    assert_eq!(crates.len(), 1);
-    assert_eq!(crates[0].name, "new-krate");
-    assert_eq!(crates[0].vers, "1.0.0");
-    assert_eq!(crates[0].deps.len(), 1);
-    assert_eq!(crates[0].deps[0].name, "my-name");
-    assert_eq!(crates[0].deps[0].package.as_ref().unwrap(), "package-name");
-}
-
-#[test]
-fn new_krate_with_dependency() {
-    let (app, _, user, token) = TestApp::full().with_token();
-
-    app.db(|conn| {
-        // Insert a crate directly into the database so that new_dep can depend on it
-        // The name choice of `foo-dep` is important! It has the property of
-        // name != canon_crate_name(name) and is a regression test for
-        // https://github.com/rust-lang/crates.io/issues/651
-        CrateBuilder::new("foo-dep", user.as_model().id).expect_build(conn);
-    });
-
-    let dependency = DependencyBuilder::new("foo-dep");
-
-    let crate_to_publish = PublishBuilder::new("new_dep")
-        .version("1.0.0")
-        .dependency(dependency);
-    token.enqueue_publish(crate_to_publish).good();
-}
-
-#[test]
-fn reject_new_krate_with_non_exact_dependency() {
-    let (app, _, user, token) = TestApp::init().with_token();
-
-    app.db(|conn| {
-        CrateBuilder::new("foo-dep", user.as_model().id).expect_build(conn);
-    });
-
-    // Use non-exact name for the dependency
-    let dependency = DependencyBuilder::new("foo_dep");
-
-    let crate_to_publish = PublishBuilder::new("new_dep")
-        .version("1.0.0")
-        .dependency(dependency);
-    token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-}
-
-#[test]
-fn new_crate_allow_empty_alternative_registry_dependency() {
-    let (app, _, user, token) = TestApp::full().with_token();
-
-    app.db(|conn| {
-        CrateBuilder::new("foo-dep", user.as_model().id).expect_build(conn);
-    });
-
-    let dependency = DependencyBuilder::new("foo-dep").registry("");
-    let crate_to_publish = PublishBuilder::new("foo").dependency(dependency);
-    token.enqueue_publish(crate_to_publish).good();
-}
-
-#[test]
-fn reject_new_crate_with_alternative_registry_dependency() {
-    let (_, _, _, token) = TestApp::init().with_token();
-
-    let dependency =
-        DependencyBuilder::new("dep").registry("https://server.example/path/to/registry");
-
-    let crate_to_publish = PublishBuilder::new("depends-on-alt-registry").dependency(dependency);
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("Cross-registry dependencies are not permitted on crates.io."),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_with_wildcard_dependency() {
-    let (app, _, user, token) = TestApp::init().with_token();
-
-    app.db(|conn| {
-        // Insert a crate directly into the database so that new_wild can depend on it
-        CrateBuilder::new("foo_wild", user.as_model().id).expect_build(conn);
-    });
-
-    let dependency = DependencyBuilder::new("foo_wild").version_req("*");
-
-    let crate_to_publish = PublishBuilder::new("new_wild")
-        .version("1.0.0")
-        .dependency(dependency);
-
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-    assert!(
-        json.errors[0].detail.contains("dependency constraints"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_twice() {
-    let (app, _, user, token) = TestApp::full().with_token();
-
-    app.db(|conn| {
-        // Insert a crate directly into the database and then we'll try to publish another version
-        CrateBuilder::new("foo_twice", user.as_model().id).expect_build(conn);
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo_twice")
-        .version("2.0.0")
-        .description("2.0.0 description");
-    let json = token.enqueue_publish(crate_to_publish).good();
-
-    assert_eq!(json.krate.name, "foo_twice");
-    assert_eq!(json.krate.description.unwrap(), "2.0.0 description");
-}
-
-#[test]
-fn new_krate_wrong_user() {
-    let (app, _, user) = TestApp::init().with_user();
-
-    app.db(|conn| {
-        // Create the foo_wrong crate with one user
-        CrateBuilder::new("foo_wrong", user.as_model().id).expect_build(conn);
-    });
-
-    // Then try to publish with a different user
-    let another_user = app.db_new_user("another").db_new_token("bar");
-    let crate_to_publish = PublishBuilder::new("foo_wrong").version("2.0.0");
-
-    let json = another_user
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("this crate exists but you don't seem to be an owner."),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_too_big() {
-    let (_, _, user) = TestApp::init().with_user();
-
-    let files = [("foo_big-1.0.0/big", &[b'a'; 2000] as &[_])];
-    let builder = PublishBuilder::new("foo_big").files(&files);
-
-    let json = user
-        .enqueue_publish(builder)
-        .bad_with_status(StatusCode::OK);
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("uploaded tarball is malformed or too large when decompressed"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_too_big_but_whitelisted() {
-    let (app, _, user, token) = TestApp::full().with_token();
-
-    app.db(|conn| {
-        CrateBuilder::new("foo_whitelist", user.as_model().id)
-            .max_upload_size(2_000_000)
-            .expect_build(conn);
-    });
-
-    let files = [("foo_whitelist-1.1.0/big", &[b'a'; 2000] as &[_])];
-    let crate_to_publish = PublishBuilder::new("foo_whitelist")
-        .version("1.1.0")
-        .files(&files);
-
-    token.enqueue_publish(crate_to_publish).good();
-}
-
-#[test]
-fn new_krate_wrong_files() {
-    let (_, _, user) = TestApp::init().with_user();
-    let data: &[u8] = &[1];
-    let files = [("foo-1.0.0/a", data), ("bar-1.0.0/a", data)];
-    let builder = PublishBuilder::new("foo").files(&files);
-
-    let json = user
-        .enqueue_publish(builder)
-        .bad_with_status(StatusCode::OK);
-    assert!(
-        json.errors[0].detail.contains("invalid tarball uploaded"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_gzip_bomb() {
-    let (_, _, _, token) = TestApp::init().with_token();
-
-    let len = 512 * 1024;
-    let mut body = io::repeat(0).take(len);
-
-    let crate_to_publish = PublishBuilder::new("foo")
-        .version("1.1.0")
-        .files_with_io(&mut [("foo-1.1.0/a", &mut body, len)]);
-
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("too large when decompressed"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_duplicate_version() {
-    let (app, _, user, token) = TestApp::init().with_token();
-
-    app.db(|conn| {
-        // Insert a crate directly into the database and then we'll try to publish the same version
-        CrateBuilder::new("foo_dupe", user.as_model().id)
-            .version("1.0.0")
-            .expect_build(conn);
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo_dupe").version("1.0.0");
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    assert!(
-        json.errors[0].detail.contains("already uploaded"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_crate_similar_name() {
-    let (app, _, user, token) = TestApp::init().with_token();
-
-    app.db(|conn| {
-        CrateBuilder::new("Foo_similar", user.as_model().id)
-            .version("1.0.0")
-            .expect_build(conn);
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo_similar").version("1.1.0");
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    assert!(
-        json.errors[0].detail.contains("previously named"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_crate_similar_name_hyphen() {
-    let (app, _, user, token) = TestApp::init().with_token();
-
-    app.db(|conn| {
-        CrateBuilder::new("foo_bar_hyphen", user.as_model().id)
-            .version("1.0.0")
-            .expect_build(conn);
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo-bar-hyphen").version("1.1.0");
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    assert!(
-        json.errors[0].detail.contains("previously named"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_crate_similar_name_underscore() {
-    let (app, _, user, token) = TestApp::init().with_token();
-
-    app.db(|conn| {
-        CrateBuilder::new("foo-bar-underscore", user.as_model().id)
-            .version("1.0.0")
-            .expect_build(conn);
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo_bar_underscore").version("1.1.0");
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    assert!(
-        json.errors[0].detail.contains("previously named"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_git_upload() {
-    let (app, _, _, token) = TestApp::full().with_token();
-
-    let crate_to_publish = PublishBuilder::new("fgt");
-    token.enqueue_publish(crate_to_publish).good();
-    app.run_pending_background_jobs();
-
-    let crates = app.crates_from_index_head("3/f/fgt");
-    assert_eq!(crates.len(), 1);
-    assert_eq!(crates[0].name, "fgt");
-    assert_eq!(crates[0].vers, "1.0.0");
-    assert!(crates[0].deps.is_empty());
-    assert_eq!(
-        crates[0].cksum,
-        "acb5604b126ac894c1eb11c4575bf2072fea61232a888e453770c79d7ed56419"
-    );
-}
-
-#[test]
-fn new_krate_git_upload_appends() {
-    let (app, _, _, token) = TestApp::full().with_token();
-
-    let crate_to_publish = PublishBuilder::new("FPP").version("0.0.1");
-    token.enqueue_publish(crate_to_publish).good();
-    let crate_to_publish = PublishBuilder::new("FPP").version("1.0.0");
-    token.enqueue_publish(crate_to_publish).good();
-    app.run_pending_background_jobs();
-
-    let crates = app.crates_from_index_head("3/f/fpp");
-    assert!(crates.len() == 2);
-    assert_eq!(crates[0].name, "FPP");
-    assert_eq!(crates[0].vers, "0.0.1");
-    assert!(crates[0].deps.is_empty());
-    assert_eq!(crates[1].name, "FPP");
-    assert_eq!(crates[1].vers, "1.0.0");
-    assert!(crates[1].deps.is_empty());
-}
-
-#[test]
-fn new_krate_git_upload_with_conflicts() {
-    let (app, _, _, token) = TestApp::full().with_token();
-
-    let index = app.upstream_repository();
-    let target = index.head().unwrap().target().unwrap();
-    let sig = index.signature().unwrap();
-    let parent = index.find_commit(target).unwrap();
-    let tree = index.find_tree(parent.tree_id()).unwrap();
-    index
-        .commit(Some("HEAD"), &sig, &sig, "empty commit", &tree, &[&parent])
-        .unwrap();
-
-    let crate_to_publish = PublishBuilder::new("foo_conflicts");
-    token.enqueue_publish(crate_to_publish).good();
-}
-
-#[test]
-fn new_krate_dependency_missing() {
-    let (_, _, _, token) = TestApp::init().with_token();
-
-    // Deliberately not inserting this crate in the database to test behavior when a dependency
-    // doesn't exist!
-    let dependency = DependencyBuilder::new("bar_missing");
-    let crate_to_publish = PublishBuilder::new("foo_missing").dependency(dependency);
-
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("no known crate named `bar_missing`"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_with_readme() {
-    let (_, _, _, token) = TestApp::full().with_token();
-
-    let crate_to_publish = PublishBuilder::new("foo_readme").readme("");
-    let json = token.enqueue_publish(crate_to_publish).good();
-
-    assert_eq!(json.krate.name, "foo_readme");
-    assert_eq!(json.krate.max_version, "1.0.0");
-}
-
-#[test]
-fn new_krate_without_any_email_fails() {
-    let (app, _, _, token) = TestApp::init().with_token();
-
-    app.db(|conn| {
-        delete(emails::table).execute(conn).unwrap();
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo_no_email");
-
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("A verified email address is required to publish crates to crates.io"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_with_unverified_email_fails() {
-    let (app, _, _, token) = TestApp::init().with_token();
-
-    app.db(|conn| {
-        update(emails::table)
-            .set((emails::verified.eq(false),))
-            .execute(conn)
-            .unwrap();
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo_unverified_email");
-
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("A verified email address is required to publish crates to crates.io"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn new_krate_records_verified_email() {
-    let (app, _, _, token) = TestApp::full().with_token();
-
-    let crate_to_publish = PublishBuilder::new("foo_verified_email");
-
-    token.enqueue_publish(crate_to_publish).good();
-
-    app.db(|conn| {
-        let email: String = versions_published_by::table
-            .select(versions_published_by::email)
-            .first(conn)
-            .unwrap();
-        assert_eq!(email, "something@example.com");
-    });
-}
-
-#[test]
 fn summary_doesnt_die() {
     let (_, anon) = TestApp::init().empty();
     anon.get::<SummaryResponse>("/api/v1/summary").good();
@@ -1727,28 +1103,6 @@ fn publish_after_yank_max_version() {
 }
 
 #[test]
-fn publish_records_an_audit_action() {
-    use cargo_registry::models::VersionOwnerAction;
-
-    let (app, anon, _, token) = TestApp::full().with_token();
-
-    app.db(|conn| assert!(VersionOwnerAction::all(&conn).unwrap().is_empty()));
-
-    // Upload a new crate, putting it in the git index
-    let crate_to_publish = PublishBuilder::new("fyk");
-    token.enqueue_publish(crate_to_publish).good();
-
-    // Make sure it has one publish audit action
-    let json = anon.show_version("fyk", "1.0.0");
-    let actions = json.version.audit_actions;
-
-    assert_eq!(actions.len(), 1);
-    let action = &actions[0];
-    assert_eq!(action.action, "publish");
-    assert_eq!(action.user.id, token.as_model().user_id);
-}
-
-#[test]
 fn yank_records_an_audit_action() {
     let (_, anon, _, token) = TestApp::full().with_token();
 
@@ -1791,158 +1145,6 @@ fn unyank_records_an_audit_action() {
     let action = &actions[2];
     assert_eq!(action.action, "unyank");
     assert_eq!(action.user.id, token.as_model().user_id);
-}
-
-#[test]
-fn publish_after_removing_documentation() {
-    let (app, anon, user, token) = TestApp::full().with_token();
-    let user = user.as_model();
-
-    // 1. Start with a crate with no documentation
-    app.db(|conn| {
-        CrateBuilder::new("docscrate", user.id)
-            .version("0.2.0")
-            .expect_build(conn);
-    });
-
-    // Verify that crates start without any documentation so the next assertion can *prove*
-    // that it was the one that added the documentation
-    let json = anon.show_crate("docscrate");
-    assert_eq!(json.krate.documentation, None);
-
-    // 2. Add documentation
-    let crate_to_publish = PublishBuilder::new("docscrate")
-        .version("0.2.1")
-        .documentation("http://foo.rs");
-    let json = token.enqueue_publish(crate_to_publish).good();
-    assert_eq!(json.krate.documentation, Some("http://foo.rs".to_owned()));
-
-    // Ensure latest version also has the same documentation
-    let json = anon.show_crate("docscrate");
-    assert_eq!(json.krate.documentation, Some("http://foo.rs".to_owned()));
-
-    // 3. Remove the documentation
-    let crate_to_publish = PublishBuilder::new("docscrate").version("0.2.2");
-    let json = token.enqueue_publish(crate_to_publish).good();
-    assert_eq!(json.krate.documentation, None);
-
-    // Ensure latest version no longer has documentation
-    let json = anon.show_crate("docscrate");
-    assert_eq!(json.krate.documentation, None);
-}
-
-#[test]
-fn bad_keywords() {
-    let (_, _, _, token) = TestApp::init().with_token();
-    let crate_to_publish =
-        PublishBuilder::new("foo_bad_key").keyword("super-long-keyword-name-oh-no");
-    token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    let crate_to_publish = PublishBuilder::new("foo_bad_key").keyword("?@?%");
-    token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    let crate_to_publish = PublishBuilder::new("foo_bad_key").keyword("áccênts");
-    token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-}
-
-#[test]
-fn good_categories() {
-    let (app, _, _, token) = TestApp::full().with_token();
-
-    app.db(|conn| {
-        new_category("Category 1", "cat1", "Category 1 crates")
-            .create_or_update(conn)
-            .unwrap();
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo_good_cat").category("cat1");
-    let json = token.enqueue_publish(crate_to_publish).good();
-
-    assert_eq!(json.krate.name, "foo_good_cat");
-    assert_eq!(json.krate.max_version, "1.0.0");
-    assert_eq!(json.warnings.invalid_categories.len(), 0);
-}
-
-#[test]
-fn ignored_categories() {
-    let (_, _, _, token) = TestApp::full().with_token();
-
-    let crate_to_publish = PublishBuilder::new("foo_ignored_cat").category("bar");
-    let json = token.enqueue_publish(crate_to_publish).good();
-
-    assert_eq!(json.krate.name, "foo_ignored_cat");
-    assert_eq!(json.krate.max_version, "1.0.0");
-    assert_eq!(json.warnings.invalid_categories, vec!["bar"]);
-}
-
-#[test]
-fn good_badges() {
-    let (_, anon, _, token) = TestApp::full().with_token();
-
-    let mut badges = HashMap::new();
-    let mut badge_attributes = HashMap::new();
-    badge_attributes.insert(
-        String::from("repository"),
-        String::from("rust-lang/crates.io"),
-    );
-    badges.insert(String::from("travis-ci"), badge_attributes);
-
-    let crate_to_publish = PublishBuilder::new("foobadger").badges(badges);
-
-    let json = token.enqueue_publish(crate_to_publish).good();
-    assert_eq!(json.krate.name, "foobadger");
-    assert_eq!(json.krate.max_version, "1.0.0");
-
-    let json = anon.show_crate("foobadger");
-    let badges = json.krate.badges.unwrap();
-    assert_eq!(badges.len(), 1);
-    assert_eq!(badges[0].badge_type, "travis-ci");
-    assert_eq!(
-        badges[0].attributes["repository"],
-        Some(String::from("rust-lang/crates.io"))
-    );
-}
-
-#[test]
-fn ignored_badges() {
-    let (_, anon, _, token) = TestApp::full().with_token();
-
-    let mut badges = HashMap::new();
-
-    // Known badge type, missing required repository attribute
-    let mut badge_attributes = HashMap::new();
-    badge_attributes.insert(String::from("branch"), String::from("master"));
-    badges.insert(String::from("travis-ci"), badge_attributes);
-
-    // Unknown badge type
-    let mut unknown_badge_attributes = HashMap::new();
-    unknown_badge_attributes.insert(String::from("repository"), String::from("rust-lang/rust"));
-    badges.insert(String::from("not-a-badge"), unknown_badge_attributes);
-
-    let crate_to_publish = PublishBuilder::new("foo_ignored_badge").badges(badges);
-
-    let json = token.enqueue_publish(crate_to_publish).good();
-    assert_eq!(json.krate.name, "foo_ignored_badge");
-    assert_eq!(json.krate.max_version, "1.0.0");
-    assert_eq!(json.warnings.invalid_badges.len(), 2);
-    assert!(json
-        .warnings
-        .invalid_badges
-        .contains(&"travis-ci".to_string(),));
-    assert!(json
-        .warnings
-        .invalid_badges
-        .contains(&"not-a-badge".to_string(),));
-
-    let json = anon.show_crate("foo_ignored_badge");
-    let badges = json.krate.badges.unwrap();
-    assert_eq!(badges.len(), 0);
 }
 
 #[test]
@@ -2138,62 +1340,6 @@ fn reverse_dependencies_query_supports_u64_version_number_parts() {
     assert_eq!(deps.versions[0].num, large_but_valid_version_number);
 }
 
-#[test]
-fn author_license_and_description_required() {
-    let (_, _, _, token) = TestApp::init().with_token();
-
-    let crate_to_publish = PublishBuilder::new("foo_metadata")
-        .version("1.1.0")
-        .unset_license()
-        .unset_description()
-        .unset_authors();
-
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-    assert!(
-        json.errors[0].detail.contains("author")
-            && json.errors[0].detail.contains("description")
-            && json.errors[0].detail.contains("license"),
-        "{:?}",
-        json.errors
-    );
-
-    let crate_to_publish = PublishBuilder::new("foo_metadata")
-        .version("1.1.0")
-        .unset_description()
-        .unset_authors()
-        .author("");
-
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-    assert!(
-        json.errors[0].detail.contains("author")
-            && json.errors[0].detail.contains("description")
-            && !json.errors[0].detail.contains("license"),
-        "{:?}",
-        json.errors
-    );
-
-    let crate_to_publish = PublishBuilder::new("foo_metadata")
-        .version("1.1.0")
-        .unset_license()
-        .license_file("foo")
-        .unset_description();
-
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-    assert!(
-        !json.errors[0].detail.contains("author")
-            && json.errors[0].detail.contains("description")
-            && !json.errors[0].detail.contains("license"),
-        "{:?}",
-        json.errors
-    );
-}
-
 /*  Given two crates, one with downloads less than 90 days ago, the
     other with all downloads greater than 90 days ago, check that
     the order returned is by recent downloads, descending. Check
@@ -2382,83 +1528,6 @@ fn test_cargo_invite_owners() {
         json.msg,
         "user cilantro has been invited to be an owner of crate guacamole"
     )
-}
-
-#[test]
-fn new_krate_tarball_with_hard_links() {
-    let (_, _, _, token) = TestApp::init().with_token();
-
-    let mut tarball = Vec::new();
-    {
-        let mut ar = tar::Builder::new(GzEncoder::new(&mut tarball, Compression::default()));
-        let mut header = tar::Header::new_gnu();
-        assert_ok!(header.set_path("foo-1.1.0/bar"));
-        header.set_size(0);
-        header.set_cksum();
-        header.set_entry_type(tar::EntryType::hard_link());
-        assert_ok!(header.set_link_name("foo-1.1.0/another"));
-        assert_ok!(ar.append(&header, &[][..]));
-        assert_ok!(ar.finish());
-    }
-
-    let crate_to_publish = PublishBuilder::new("foo").version("1.1.0").tarball(tarball);
-
-    let json = token
-        .enqueue_publish(crate_to_publish)
-        .bad_with_status(StatusCode::OK);
-
-    assert!(
-        json.errors[0]
-            .detail
-            .contains("too large when decompressed"),
-        "{:?}",
-        json.errors
-    );
-}
-
-#[test]
-fn publish_new_crate_rate_limited() {
-    let (app, anon, _, token) = TestApp::full()
-        .with_publish_rate_limit(Duration::from_millis(500), 1)
-        .with_token();
-
-    // Upload a new crate
-    let crate_to_publish = PublishBuilder::new("rate_limited1");
-    token.enqueue_publish(crate_to_publish).good();
-
-    // Uploading a second crate is limited
-    let crate_to_publish = PublishBuilder::new("rate_limited2");
-    token
-        .enqueue_publish(crate_to_publish)
-        .assert_status(StatusCode::TOO_MANY_REQUESTS);
-    app.run_pending_background_jobs();
-
-    anon.get::<()>("/api/v1/crates/rate_limited2")
-        .assert_status(StatusCode::NOT_FOUND);
-
-    // Wait for the limit to be up
-    thread::sleep(Duration::from_millis(500));
-
-    let crate_to_publish = PublishBuilder::new("rate_limited2");
-    token.enqueue_publish(crate_to_publish).good();
-
-    let json = anon.show_crate("rate_limited2");
-    assert_eq!(json.krate.max_version, "1.0.0");
-}
-
-#[test]
-fn publish_rate_limit_doesnt_affect_existing_crates() {
-    let (app, _, _, token) = TestApp::full()
-        .with_publish_rate_limit(Duration::from_millis(500), 1)
-        .with_token();
-
-    // Upload a new crate
-    let crate_to_publish = PublishBuilder::new("rate_limited1");
-    token.enqueue_publish(crate_to_publish).good();
-
-    let new_version = PublishBuilder::new("rate_limited1").version("1.0.1");
-    token.enqueue_publish(new_version).good();
-    app.run_pending_background_jobs();
 }
 
 #[test]
